+++ etc/bin/run-tests.sh
@@ -8,6 +8,9 @@ function do_tests () {
     1)
       ant test-generative 2>&1 | tee test-generative.log
       ;;
+    2)
+      bash etc/bin/test-cider.sh 2>&1 | tee test-cider.log
+      ;;
   esac
 }
 
@@ -21,19 +24,29 @@ else
   ( export CIRCLE_NODE_INDEX=0;
     do_tests > /dev/null ) &
   tex=$1
+
   ( export CIRCLE_NODE_INDEX=1;
     do_tests > /dev/null ) &
   teg=$1
+
+  ( export CIRCLE_NODE_INDEX=2;
+    do_tests > /dev/null ) &
+  tec=$1
+  
   wait $tex
   texr=$?
   wait $teg
   tegr=$?
-  if [ $texr ] && [ $tegr ]
+  wait $tec
+  tecr=$?
+
+  if [ $texr ] && [ $tegr ] && [ $tecr ]
   then
     echo "Tests OK!"
   else
     [ ! $texr ] && echo "Example tests failed, see test-example.log"
     [ ! $tegr ] && echo "Generative tests failed, see test-generative.log"
+    [ ! $tecr ] && echo "CIDER tests failed, see test-cider.log"
     echo "Tests failed :("
     exit 1
   fi